[FLINK-3511] [hadoop-compatibility] Move hadoop-compatibility examples to test scope
